+++ src/openai-service.ts
@@ -75,90 +75,6 @@ export class OpenAIService {
     return Math.ceil(text.length / 4);
   }
 
-  private transformToDuckAIRequest(
-    request: ChatCompletionRequest,
-  ): DuckAIRequest {
-    // Use the model from request, fallback to default
-    const model = request.model || "gpt-4o-mini";
-
-    if (!(model in DUCKAI_MODELS)) {
-      throw new Error(
-        `Model ${model} is not a valid model, valid models: ${Object.keys(DUCKAI_MODELS).join(", ")}`,
-      );
-    }
-
-    const transformedMessages: DuckChatCompletionMessage[] = [];
-
-    for (const message of request.messages as ChatCompletionMessage[]) {
-      if (Array.isArray(message.content)) {
-        const transformedContent = [];
-        for (const content of message.content as ChatCompletionContentPart[]) {
-          if (content.type == "text") {
-            if (typeof content.text !== "string" || content.type !== "text") {
-              throw new Error("Image text must be a string of type text");
-            } else {
-              transformedContent.push(content);
-            }
-          } else if (content.type == "image_url") {
-            if (
-              content.image_url === null ||
-              typeof content.image_url?.url !== "string" ||
-              content.type !== "image_url"
-            ) {
-              throw new Error("Image payload is incorrect");
-            } else {
-              // valid image, transform to DuckChatCompletionRequest
-              const newImagePayload: DuckChatCompletionContentPartImage = {
-                image: content.image_url.url,
-                type: "image",
-                mimeType: content.image_url.url
-                  .split(",")[0]
-                  .split(":")[1]
-                  .split(";")[0],
-              };
-
-              console.log(newImagePayload);
-
-              transformedContent.push(newImagePayload);
-            }
-          }
-        }
-
-        // transform message
-        const clonedMessage = structuredClone(message);
-        const newMessage = {
-          ...clonedMessage,
-          content: transformedContent,
-        };
-        transformedMessages.push(newMessage as DuckChatCompletionMessage);
-      } else {
-        transformedMessages.push(message as DuckChatCompletionMessage);
-      }
-    }
-
-    // validate reasoning effort
-    const reasoning_effort =
-      request.reasoning_effort || DUCKAI_MODELS[model].reasoning_effort;
-
-    if (
-      DUCKAI_MODELS[model].valid_reasoning_efforts != undefined &&
-      !DUCKAI_MODELS[model].valid_reasoning_efforts?.includes(reasoning_effort)
-    ) {
-      throw new Error(
-        `Model ${model} does not support this reasoning effort (${reasoning_effort}),
-        valid reasoning efforts: ${(DUCKAI_MODELS[model].valid_reasoning_efforts || []).join(", ")}`,
-      );
-    }
-
-    return {
-      canUseTools: true,
-      messages: transformedMessages,
-      metadata: request.metadata,
-      model,
-      reasoningEffort: reasoning_effort,
-    };
-  }
-
   async createChatCompletion(
     request: ChatCompletionRequest,
   ): Promise<ChatCompletionResponse> {
@@ -172,7 +88,7 @@ export class OpenAIService {
       return this.createChatCompletionWithTools(request);
     }
 
-    const duckAIRequest = this.transformToDuckAIRequest(request);
+    const duckAIRequest = DuckAI.transformToDuckAIRequest(request);
     const response = await this.duckAI.chat(duckAIRequest);
 
     const id = this.generateId();
@@ -237,7 +153,7 @@ Please follow these instructions when responding to the following user message.`
       });
     }
 
-    const duckAIRequest = this.transformToDuckAIRequest({
+    const duckAIRequest = DuckAI.transformToDuckAIRequest({
       ...request,
       messages: modifiedMessages,
     });
@@ -428,7 +344,7 @@ Please follow these instructions when responding to the following user message.`
       return this.createChatCompletionStreamWithTools(request);
     }
 
-    const duckAIRequest = this.transformToDuckAIRequest(request);
+    const duckAIRequest = DuckAI.transformToDuckAIRequest(request);
     const duckStream = await this.duckAI.chatStream(duckAIRequest);
 
     const id = this.generateId();
